requirements.txt
argh==0.26.2
asn1crypto==0.24.0
attrs==18.1.0
backpack==0.1
bcrypt==3.1.4
beautifulsoup4==4.6.0
blinker==1.4
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
cleo==0.6.5
click==6.7
cryptography==2.2.2
Faker==0.7.18
idna==2.6
inflection==0.3.1
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
masonite==2.0.0
masonite-entry==0.1.6
orator==0.9.7
pafy==0.5.4
passlib==1.7.1
pastel==0.1.0
pathtools==0.1.2
pendulum==1.4.4
pluggy==0.6.0
psycopg2==2.7.3.2
py==1.5.3
pyaml==16.12.2
pycparser==2.18
Pygments==2.2.0
pylev==1.3.0
PyMySQL==0.8.0
pytest==3.3.1
python-dateutil==2.7.3
python-dotenv==0.7.1
python-vlc==3.0.102
pytz==2018.4
pytzdata==2018.5
PyYAML==3.12
requests==2.18.4
requests-file==1.4.3
simplejson==3.15.0
six==1.11.0
tabulate==0.8.2
tldextract==2.2.0
transitions==0.6.8
triggers==0.0.3
tzlocal==1.5.1
urllib3==1.22
validator.py==1.2.5
waitress==1.1.0
watchdog==0.8.3
whitenoise==3.3.1
wrapt==1.10.11
youtube-dl==2018.8.22
